Add real VMU save to control center. Modify vmu_init. Now there is a lot of function available here for vmu stuff.

+++ lua/vmu_init.lua
@@ -1,7 +1,7 @@
 --- @date 2002/12/06
 --- @author benjamin gerard <ben@sashipa.com>
 --- @brief  LUA script to initialize dcplaya VMU backup.
---- $Id: vmu_init.lua,v 1.14 2003-03-06 12:38:30 ben Exp $
+--- $Id: vmu_init.lua,v 1.15 2003-03-08 18:30:44 ben Exp $
 ---
 
 -- Unload library
@@ -9,10 +9,30 @@ vmu_init_loaded = nil
 
 --- Global VMU path variable
 vmu_path = nil
+vmu_leaf = nil
+ram_path = ram_path or nil
 
-function ramdisk_init()
+--- Get current VMU file.
+function vmu_file(path,leaf)
+   path = path or vmu_path
+   leaf = leaf or vmu_leaf
+   return type(path) == "string"
+      and type(leaf) == "string"
+      and canonical_path(path .. '/' .. leaf)
+end
+
+--- Set current VMU file.
+function vmu_set_file(filepath)
+   vmu_path,vmu_leaf = get_path_and_leaf(filepath)
+   return vmu_file()
+end
+
+function ramdisk_init(path)
    print("Initializing RAM disk.")
-   if not test ("-d","/ram") then
+   ram_path = path or ram_path or "/ram"
+
+   if not test ("-d",ram_path) then
+      ram_path = nil
       print("ramdisk_init : no ramdisk")
       return
    end
@@ -21,12 +41,9 @@ function ramdisk_init()
    for i,v in { "tmp", "dcplaya",
       "dcplaya/lua", "dcplaya/configs", "dcplaya/playlists" } do
       print("ramdisk_init : create [" .. v .. "] directory")
-      mkdir("/ram/" .. v)
+      mkdir(ram_path .. "/" .. v)
    end
 end
-
--- Initialze ram disk has soon as possible
-ramdisk_init()
 
 -- Load required libraries
 if not dolib("vmu_select") then return nil end
@@ -46,25 +63,33 @@ if not dolib("textviewer") then return nil end
 ---
 function vmu_save_file(fname, path)
    local errstr
-   local tmpfile="/ram/tmp/backup.dcar"
-   local headerfile="/rd/vmu_header.bin"
 
-   fname = fullpath(fname)
-   path = fullpath(path)
+   if type(ram_path) ~= "string" then
+      errstr = "no ramdisk found. Can't create temporary file."
+   else
+      local tmpfile = ram_path .. "/tmp/backup.dcar"
+      local headerfile = "/rd/vmu_header.bin"
+      fname = canonical_path(fname)
+      path = canonical_path(path)
 
-   -- Create temporary backup file
-   unlink(tmpfile)
-   if not copy("-fv", headerfile, tmpfile) then
-      errstr = "error copying header file"
-   elseif not dcar("av9", tmpfile, path) then
-      errstr = "error creating temporary archive [" 
-	 .. tmpfile .. "] from [" .. tostring(path) .. "]"
-   elseif not copy("-fv", tmpfile, fname) then
-      errstr = "error copying [" .. tmpfile ..
-	 "] to [" .. tostring(fname) .. "]"
+      if type(fname) ~= "string" or type(path)  ~= "string" then
+	 errstr = "bad parameter"
+      else
+	 -- Create temporary backup file
+	 unlink(tmpfile)
+	 if not copy("-fv", headerfile, tmpfile) then
+	    errstr = "error copying header file"
+	 elseif not dcar("av9", tmpfile, path) then
+	    errstr = "error creating temporary archive [" 
+	       .. tmpfile .. "] from [" .. tostring(path) .. "]"
+	 elseif not copy("-fv", tmpfile, fname) then
+	    errstr = "error copying [" .. tmpfile ..
+	       "] to [" .. tostring(fname) .. "]"
+	 end
+	 unlink(tmpfile)
+      end
    end
    print("vmu_save_file : " .. (errstr or "success"))
-   unlink(tmpfile)
    return errstr == nil
 end
 
@@ -88,52 +113,89 @@ function vmu_load_file(fname,path)
 end
 
 function dcplaya_welcome()
+end
 
 
+--- Get pluged VMU list.
+function vmu_list()
+   -- Check available VMU
+   local dir = dirlist("-n", "/vmu")
+   if type(dir) ~= "table" or dir.n < 1 then
+      print("vmu_list : no VMU found.")
+      return
+   end
+   return dir
+end
+
+--- Get dcplaya VMU files.
+function vmu_find_files(dir, expr)
+   expr = expr or "^dcplaya.*"
+   dir = dir or vmu_list()
+   if type(dir) ~= "table" then return end
+
+   local found = {}
+   -- Check for available dcplaya save files.
+   local i
+   for i=1, dir.n do
+      local v = dir[i]
+      local vmudir
+      local path = "/vmu/" .. v.name .. "/"
+      print("Scanning VMU .." .. path)
+      vmudir = dirlist(path)
+      if type(vmudir) == "table" then
+	 local j
+	 for j=1, vmudir.n do
+	    local w = vmudir[j]
+	    print(" -> " .. w.name)
+	    if strfind(w.name,expr) then
+	       w.path = v.name
+	       tinsert(found, w)
+	       print(" + Added " .. w.path.. "/" .. w.name)
+	    end
+	 end
+      end
+   end
+   return getn(found) > 0 and found
+end
+
+--- Choose a VMU.
+function vmu_choose()
+   local vs = vmu_select_create(nil,"Select VMS")
+   if not vs then
+      print("vmu_choose : error VMS selection")
+   else
+      local choice = evt_run_standalone(vs)
+      if type(choice) == "string" then
+	 return "/vmu/" .. choice
+      end
+   end
+end
+
+function vmu_choose_file(files, dir, expr)
+   files = files or vmu_find_files(dir, expr)
+   if type(files) ~= "table" then return end
+   local vs = vmu_select_create(nil,"Select a file", files)
+   if not vs then
+      print("vmu_choose_file : error VMS selection")
+   else
+      choice = evt_run_standalone(vs)
+      if type(choice) == "string" then
+	 return vmu_set_file("/vmu/" .. choice)
+      end
+   end
 end
 
 --- Initialise VMU path.
 --
-function vmu_init()
+function vmu_init(force_choice)
    local done
-   local force_choice -- Avoid infinite loop when only one file and corrupt
+   local choice = nil
 
    while not done do
       local dir
+      local found = vmu_find_files()
 
-      -- Check available VMU
-      dir = dirlist("-n", "/vmu")
-      if type(dir) ~= "table" or dir.n < 1 then
-	 print("vmu_init : no VMU found.")
-	 return
-      end
-
-      local found = {}
-      -- Check for available dcplaya save files.
-      local i
-      for i=1, dir.n do
-	 local v = dir[i]
-	 local vmudir
-	 local path = "/vmu/" .. v.name .. "/"
-	 print("Scanning VMU .." .. path)
-	 vmudir = dirlist(path)
-	 if type(vmudir) == "table" then
-	    local j
-	    for j=1, vmudir.n do
-	       local w = vmudir[j]
-	       print(" -> " .. w.name)
-	       if strfind(w.name,"^dcplaya.*") then
-		  w.name = v.name .. "/" .. w.name
-		  tinsert(found, w)
-		  print(" + Added " .. w.name)
-	       end
-	    end
-	 end
-      end
-
-      local choice = nil
-
-      --- Must selct a file if more than one was found
+      --- Must select a file if more than one was found
       if found.n and found.n > 1 then
 	 local result
 	 result = gui_ask('More than one dcplaya save file has been found. You are going to select the one you want to use.',
@@ -141,16 +203,15 @@ function vmu_init()
 			  nil,
 			  "More than one dcplaya file")
 	 if result and result == 1 then
-	    local vs = vmu_select_create(nil,"Select a file", found)
-	    if not vs then
-	       print("vmu_init : error VMS selection")
-	    else
-	       choice = evt_run_standalone(vs)
-	    end
+	    choice = vmu_choose_file(found)
+	 else
+	    choice = vmu_set_file(nil)
+	    return
 	 end
       else
 	 if not force_choice and found.n and found.n == 1 then
-	    choice = found[1]
+	    choice = "/vmu/" .. found[1].path .. "/" .. found[1].name
+	    printf("CHOICE UNIC %q",choice)
 	 else
 	    local ok
 	    while not ok do
@@ -195,21 +256,14 @@ function vmu_init()
       end
       
       if not choice then
-	 local vs = vmu_select_create(nil,"Select VMS")
-	 if not vs then
-	    print("vmu_init : error VMS selection")
-	 else
-	    choice = evt_run_standalone(vs)
-	    if type(choice) == "table" then
-	       vmu_path = "/vmu/" .. choice.name
-	       done = 1
-	    end
+	 choice = vmu_choose()
+	 if choice then
+	    choice = vmu_set_file(choice .. "/dcplaya.01")
 	 end
       else
-	 vmu_path = "/vmu/" .. choice.name
-	 done = vmu_load_file(vmu_path, "/ram/dcplaya")
-
+	 done = vmu_load_file(choice, "/ram/dcplaya")
 	 if not done then
+	    choice = vmu_set_file(nil)
 	    local result
 	    result = gui_ask('<left>Invalid dcplaya save file. File may be corrupt.<br><center><img name="grimley" src="stock_grimley.tga" scale="1.5">',
 			     { "try another","cancel" },
@@ -225,8 +279,27 @@ function vmu_init()
 	 end
       end
    end
+   printf("VMU [ path=%q leaf=%q path=%q ]",
+	  tostring(vmu_path),tostring(vmu_leaf),
+	  tostring(vmu_file()))
 end
 
-vmu_init()
-print("VMU path [" .. tostring(vmu_path) .. "]")
+function vmu_confirm_write(path)
+   local r
+   if type(path) ~= "string" then
+      return
+   end
 
+   local col = color_tostring(gui_text_color) or "#FFFFB0"
+   r = gui_yesno(
+'<macro macro-name="red" macro-cmd="font" color="#FF0000" size="18">' ..
+'<macro macro-name="green" macro-cmd="font" color="#00FF00" size="18">' ..
+'<macro macro-name="nrm" macro-cmd="font" color="'..col..'" size="16">' ..
+'<nrm>You are going to write data on your VMU.' ..
+'<p><vspace h="8"><green>['..path..']'..
+'<p><vspace h="8"><left><nrm>This is a very <red>experimental<nrm> function of dcplaya. You may <red>lost<nrm> all data and/or <red>damage<nrm> your VMU.<p><vspace h="8">If you confirm <red>do not switch off<nrm> your dreamcast during the process or you will <red>lost<nrm> everything !!<br><center>', 400, "Confirm VMU write", "Write the VMU", "cancel")
+   return r and r == 1;
+end
+
+vmu_init_loaded = 1
+return vmu_init_loaded
